requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pandas==2.0.3
numpy==1.24.3  # Add this line to pin numpy
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
gdown==5.2.0
fuzzywuzzy
asyncpg